# openapi/finanzas.yaml
openapi: 3.1.0
info:
  title: Finanzas API - Financial Planning & Service Delivery Control
  version: 1.0.0
  description: |
    Complete API for managing financial planning, project tracking, payroll integration,
    budget allocations, and service delivery control. Supports prefactura webhook integration
    and comprehensive budget tracking for SDT (Service Delivery Team).

    **Region**: us-east-2  
    **Base Path**: /finanzas/  
    **Authentication**: OIDC via AWS Cognito JWT  
    **RBAC**: JWT groups include `SDT` for Service Delivery Team access
  contact:
    name: API Support
    email: support@finanzas.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://{apiId}.execute-api.us-east-2.amazonaws.com/{stage}/finanzas
    description: AWS API Gateway endpoint (us-east-2)
    variables:
      apiId:
        default: your-api-id
        description: API Gateway ID (replace with actual ID after deployment)
      stage:
        default: dev
        description: Deployment stage (dev, staging, prod)
  - url: http://localhost:3000/finanzas
    description: Local Development Server
...
security:
  - CognitoJWT: []
paths:
...
  /projects/{id}/allocations:bulk:
    put:
      tags:
        - Allocations
      summary: Bulk update allocations
      description: Updates budget allocations for multiple rubros in a single operation. Supports both planned and forecast allocations via type parameter.
      operationId: bulkUpdateAllocations
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
            pattern: "^proj_[a-z0-9]{10}$"
        - name: type
          in: query
          required: false
          description: Type of allocation to update (planned or forecast). Defaults to planned.
          schema:
            type: string
            enum: [planned, forecast]
            default: planned
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllocationBulk"
            example:
              allocations:
                - rubro_id: "rubro_1a2b3c4d5e"
                  mes: "2025-11"
                  monto_planeado: 30000
                - rubro_id: "rubro_1a2b3c4d5e"
                  mes: "2025-12"
                  monto_planeado: 32000
                - rubro_id: "rubro_6f7g8h9i0j"
                  mes: "2025-11"
                  monto_planeado: 15000
      responses:
        "200":
          description: Allocations updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated_count:
                    type: integer
                    description: Number of allocations updated
                  allocations:
                    type: array
                    items:
                      type: object
                      properties:
                        rubro_id:
                          type: string
                        mes:
                          type: string
                        monto_planeado:
                          type: number
                        status:
                          type: string
              example:
                updated_count: 3
                allocations:
                  - rubro_id: "rubro_1a2b3c4d5e"
                    mes: "2025-11"
                    monto_planeado: 30000
                    status: "updated"
                  - rubro_id: "rubro_1a2b3c4d5e"
                    mes: "2025-12"
                    monto_planeado: 32000
                    status: "updated"
                  - rubro_id: "rubro_6f7g8h9i0j"
                    mes: "2025-11"
                    monto_planeado: 15000
                    status: "updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /budgets/all-in:
    get:
      tags:
        - Hub
      summary: Get Annual All-In Budget
      description: |
        Returns the organization-wide annual all-in budget for a specific year.
        Only accessible to SDMT and EXEC_RO roles.
      operationId: getAnnualBudget
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            minimum: 2020
            maximum: 2100
          description: Year for the budget (e.g., 2025)
      responses:
        "200":
          description: Annual budget retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnualBudget"
              example:
                year: 2025
                amount: 5000000
                currency: USD
                lastUpdated: "2025-01-15T10:30:00Z"
                updatedBy: "user@example.com"
        "404":
          description: No budget found for the specified year
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "No budget found for year 2025"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags:
        - Hub
      summary: Set Annual All-In Budget
      description: |
        Sets or updates the organization-wide annual all-in budget for a specific year.
        Only accessible to SDMT role.
      operationId: setAnnualBudget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnualBudgetInput"
            example:
              year: 2025
              amount: 5000000
              currency: USD
      responses:
        "200":
          description: Annual budget set successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnualBudget"
              example:
                year: 2025
                amount: 5000000
                currency: USD
                lastUpdated: "2025-01-15T10:30:00Z"
                updatedBy: "user@example.com"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    CognitoJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        AWS Cognito JWT token. Include the token in the Authorization header as: 
        `Authorization: Bearer <token>`

        Token must include valid groups claim with 'SDT' group for Service Delivery Team access.
  schemas:
...
    AllocationBulk:
      type: object
      required:
        - allocations
      properties:
        allocations:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - rubro_id
              - mes
            properties:
              rubro_id:
                type: string
                pattern: "^rubro_[a-z0-9]{10}$"
              mes:
                type: string
                pattern: '^\d{4}-\d{2}$'
              monto_planeado:
                type: number
                minimum: 0
                description: Planned amount (used when type=planned)
              monto_proyectado:
                type: number
                minimum: 0
                description: Forecast amount (used when type=forecast)
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code for programmatic handling
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
    AnnualBudget:
      type: object
      required:
        - year
        - amount
        - currency
      properties:
        year:
          type: integer
          description: Budget year
        amount:
          type: number
          minimum: 0
          description: Total annual budget amount
        currency:
          type: string
          enum: [USD, EUR, MXN]
          description: Budget currency
        lastUpdated:
          type: string
          format: date-time
          description: Last update timestamp
        updatedBy:
          type: string
          description: User who last updated the budget
    AnnualBudgetInput:
      type: object
      required:
        - year
        - amount
      properties:
        year:
          type: integer
          minimum: 2020
          maximum: 2100
          description: Budget year
        amount:
          type: number
          minimum: 0
          description: Total annual budget amount
        currency:
          type: string
          enum: [USD, EUR, MXN]
          default: USD
          description: Budget currency
  responses:
    BadRequest:
      description: Bad Request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "BAD_REQUEST"
            message: "Invalid request parameters"
            details:
              field: "mod_total"
              error: "Must be a positive number"
            timestamp: "2025-10-31T03:55:00Z"
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "UNAUTHORIZED"
            message: "Authentication required"
            details:
              reason: "Missing or invalid JWT token"
            timestamp: "2025-10-31T03:55:00Z"
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "FORBIDDEN"
            message: "Insufficient permissions to access this resource"
            details:
              required_group: "SDT"
              user_groups: ["users"]
            timestamp: "2025-10-31T03:55:00Z"
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "NOT_FOUND"
            message: "Resource not found"
            details:
              resource_type: "project"
              resource_id: "proj_invalid123"
            timestamp: "2025-10-31T03:55:00Z"